+++ packages/ember-intl-update/src/steps/update-package/to-v8.ts
@@ -6,14 +6,13 @@ import {
   updatePackageJson,
 } from './to-v8/index.js';
 
-// eslint-disable-next-line @typescript-eslint/require-await
 export async function updatePackageToV8(options: Options): Promise<Todos> {
   const todosAll: Todos[] = [];
 
   todosAll.push(updatePackageJson(options));
   updateEmberIntlConfig(options);
   todosAll.push(updateFormatsConfig(options));
-  renameFormatRelative(options);
+  await renameFormatRelative(options);
 
   return todosAll.flat();
 }
