+++ packages/ember-intl-update/src/steps/update-package/to-v8/rename-format-releative/update-service-invocations.ts
@@ -1,0 +1,48 @@
+import { AST } from '@codemod-utils/ast-javascript';
+
+type Data = {
+  from: string;
+  to: string;
+};
+
+export function updateServiceInvocations(
+  file: string,
+  data: Data,
+): {
+  file: string;
+  update: boolean;
+} {
+  const traverse = AST.traverse(true);
+  let update = false;
+
+  const ast = traverse(file, {
+    visitMemberExpression(path) {
+      this.traverse(path);
+
+      if (
+        path.node.object.type !== 'MemberExpression' ||
+        path.node.object.property.type !== 'Identifier' ||
+        path.node.object.property.name !== 'intl'
+      ) {
+        return false;
+      }
+
+      if (
+        path.node.property.type !== 'Identifier' ||
+        path.node.property.name !== data.from
+      ) {
+        return false;
+      }
+
+      path.node.property.name = data.to;
+      update = true;
+
+      return false;
+    },
+  });
+
+  return {
+    file: AST.print(ast),
+    update,
+  };
+}
