+++ packages/ember-intl-update/src/steps/update-package/to-v8/rename-format-releative/worker.ts
@@ -1,0 +1,5 @@
+import { runWorker } from '@codemod-utils/threads';
+
+import { task } from './task.js';
+
+runWorker(task);
